refactor: separate people repository implementations from people repository interface
